LayoutTests/fast/events/touch/ios/touch-event-radius-expected.txt

PASS radiusX is expected
PASS radiusY is expected
PASS successfullyParsed is true

TEST COMPLETE
